Added organs for scenes
Other than heart (which has already been completed) and kidneys (the anatomy package lacks kidneys) I have copied and pasted the contents of the heart scene to the other scene and replaced the heart with the scene's respective organ. Also edited the text above the organs. For the brain I also resized, but the other organs didn't need it.